Re-implementation of the host tool compilation IR pass for the new parser interface.
